requirements.txt
Flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
openai==1.3.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0 